+++ CHANGELOG.md
@@ -8,3 +8,8 @@ This document describes all changes made to SilverLibs.
   * This fixes the remaining bug related to the gear locking feature.
 * Removed Kaustra from dynamic elemental belt equipping logic (Hachirin and Orpheus).
   * It was decided that when using Kaustra, a SCH will always control the weather to make Hachirin-no-Obi always the best choice. Will let them set this in their Kaustra set and not override it with SilverLibs logic.
+
+## 2025.JUL.13.1
+* Added silibs.last_midcast_set variable which stores the midcast set used in your most recent action.
+  * If you are using my RDM and BLM luas you should remove any setters for the last_midcast_set variable from there and use the silibs.last_midcast_set variable instead.
+  * This was necessary due to the change in the last patch for where and when SilverLibs action hooks are executed.